requirements.txt
streamlit == 0.83.0
pandas == 1.1.3
imageio == 2.9.0
plotly_express == 0.4.1

